test formatScore
